jobsearch-playwright-mcp-pydanticai/pyproject.toml
[project]
name = "jobsearch-playwright-mcp-pydanticai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.84.0",
    "pydantic-ai>=0.2.15",
    "pydantic-ai-slim[mcp]>=0.2.15",
    "pypdf2>=3.0.1",
]
